Allow Landlock Syscalls
Allowlisting the three Landlock syscalls, because these would allow the
Clank GPU process to run under a stronger sandbox.

There is low risk from allowlisting these, because all they allow a
caller to do is to further restrict its capabilities.

For background on the motivation here, see go/landlock-clank-design.
We don’t require a libc wrapper for this allowlisting to be useful.

Bug: http://b/399173273
Test: Verified successful Landlock syscall execution in a modified
Chrome binary

Change-Id: I5fcbffc290ab51906cd06a5994bfe68de54aa384