Catch up with a few years worth of "syscalls we should probably use".
In particular, glibc is about to expose openat2() despite the ABI concerns, so that's like to be a source compatibility issue going forward.

faccessat2()/fchmodat2() have long been on the to-do list, but with no obvious way to switch faccessat()/fchmodat() over that isn't massively painful.

The *xattrat() functions would be useful for race-free selinux operations. (Though libselinux is such bad code, I fear that's the least of its issues.)

Change-Id: Ie52f58239ec06ae261ddea7d4095d792cd2bf377