<sys/types.h>: remove ancient `ssize_t` hack.
Despite the comment, this has been in the NDK for about a decade now. The referenced bug is so old it's already been deleted, so there's no more background to this. I'm guessing the comment was strictly true, but that we already made the incompatible mangled name change by accident, and either no-one noticed or no-one said anything.

That, of course, makes me wonder how many of our "this isn't exactly the type POSIX says it should be, but fixing it could cause incompatible mangled names for C++ users" warts could similarly be removed without anyone noticing. But since the benefit is also small, there's no particular reason to "try it and see".

Change-Id: I126ab545635b20e753b31cd378cc32a84e8de544