sys_hwprobe_test.cpp: remove unnecessary __has_include().
The __has_include() at the top of the file is necessary, as are the __riscv tests, because we try to ensure all tests are built in all combinations, even where they don't make sense. In part this is for historical reasons where CTS would take a list of tests from one architecture and assume that was the correct list for all architectures, but also to make it less likely that we accidentally don't build/run tests --- by messing up the "only on <arch>" logic in the .bp file, say.

Having __has_include() checks that _aren't_ guarding a #include goes against these goals.

Change-Id: I67ea1755803db6c8a0db796998127495d5bf5398