dl_iterate_phdr(): fix iteration order.
Code that was expecting the main executable to come first was upset to
find the dynamic linker first and the main executable second.

Switch to more intention-revealing names for the various special soinfo
objects to make things a bit clearer. Most of the work of fixing this
bug was understanding what these names really meant, and which places
were accidentally using the wrong name.

Bug: http://b/397501154
Test: treehugger
Change-Id: I076a16013ee2ca91f5e6aebd024723948db5a6ca