16k compat: Check RELRO end alignment
A binary with LOAD segment alignments of at least 16KiB can still be
incompatible with 16KiB pages if the first RW segment has a RELRO prefix
ending at a non-16KiB-aligned address. If this is the case, the ELF file
should be loaded using 16KiB compatibility mode, otherwise some RW
regions will be incorrectly marked RO, causing segmentation faults.

We check for this case only if the binary appears to have been linked
with a conventional linker, i.e. it satisfies the following:
* Has a single GNU_RELRO segment,
* The GNU_RELRO segment starts at the beginning of a LOAD segment,
* The GNU_RELRO segment is entirely contained in a LOAD segment, or
  the end of the GNU_RELRO segment is at least 4KiB-aligned.

Test: manual
Bug: 389761448
Signed-off-by: Csanád Hajdú <csanad.hajdu@arm.com>
[ Move relro_min_align compat logic to linker_phdr_16kib_compat.cpp
  Fix comment: s/min_palign/min_align_
  Skip relro_min_align for non-16KiB devices
    - Kalesh Singh ]
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>
Change-Id: Ib9ba536dd3e53b42daf58c3679d242969d0dd4e6