Maximize __BEGIN_DECLS/__END_DECLS scope.
Following "don't make me think", make it (a) not matter where you add a function (as long as it's after the #includes) and (b) clear to readers that everything is covered (even things that don't need to be). This was already largely the case for newer/more rewritten headers, but consistency is good.

This came up because it's possible that the "FIXME" mentioned in https://github.com/android/ndk/issues/2161 existed because it wasn't immediately clear at a glance that the __BEGIN_DECLS/__END_DECLS in question was superfluous.

Change-Id: Ibe8722b6a8b58683703d902e28f23501ef00e638