Fix -Wtautological-type-limit-compare.
The comparison with 0xFFFFFFFF is always false because that's the
maximum value of uint32_t. Change 'parts' to unsigned long long
and use strtoull() so that the check is not tautological on 32-bit
platforms.

This seems to be the only file in Android that triggers this
warning.

Bug: 296606895
Test: presubmit
Flag: EXEMPT trivial warning fix
Change-Id: I90689830c40b8c9265c1013380d7dfb38fcd8d11