Stop using the cached pid.
The other Linux libcs (glibc and musl) stopped caching the pid because it's too error-prone and breaks some uses of clone(). Rather than have to try to paper over the same problems with clone3(), let's just stop punching ourselves in the face.

For now we leave the _tid_ caching in place even though that's also caused trouble for some advanced users.

This change only stops _using_ the cached pid, because my attempt to remove all of the pid-caching code refuses to boot, and removing pieces one at a time seems like the least worst way to debug that.

Bug: http://b/65119051
Bug: https://issuetracker.google.com/130509957
Change-Id: I89b35fcd4127752f4db7bc8d68a5657d4b79fdd4