Reimplement raise() in terms of inline_raise().
When pid caching is removed, we won't need the syscall(__NR_getpid) hack, which made me notice we had two copies of this. We haven't cared much because we're already calling inline_raise() rather than raise() internally, but we may as well make the benefits available to anyone who does call raise(), since we've written the code.

Change-Id: I52546f1f89eac55afbe982ef24991f90d1774654