strtok_r()/wcstok() tests: check the state pointer too.
This is a little bit odd since no specification says that the reason why "subsequent searches return null" is because the state pointer is null ... but that's the only existing (and only reasonable?) implementation, and _not_ checking that risks tests passing "by accident" because there happens to be another '\0' or L'\0' after the initial one.

An alternative would be to have a _loop_ checking that you can call the function a large number of times and keep getting null, so "why not both?".

Also use EXPECT_EQ() rather than EXPECT_STREQ() when we're expecting a null pointer.

Change-Id: I7f12eac78d2adda978abff6f88e1f62db59763d3